+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.0.2
 
 - Move to using [nuxt-directus package](https://nuxt-directus.netlify.app/)
+- Add @nuxtus/cli
 - Add .DS_Store to gitignore
 
 ## 0.0.1